common utility functions for c++ extensions
