+++ sktime/clustering/setup.py
@@ -1,4 +1,4 @@
-
+import os
 
 def configuration(parent_package='', top_path=None):
     from numpy.distutils.misc_util import Configuration
@@ -6,7 +6,7 @@ def configuration(parent_package='', top_path=None):
 
     config.add_extension('_clustering_bindings',
                          sources=['src/clustering_module.cpp'],
-                         include_dirs=['include'],
+                         include_dirs=['include', os.path.join(top_path, 'sktime', 'src', 'include')],
                          language='c++',
                          )
     return config
