--- sktime/markov/_bindings/include/common.h
@@ -1,15 +1,0 @@
-//
-// Created by mho on 7/18/19.
-//
-
-#pragma once
-
-#include <pybind11/pybind11.h>
-#include <pybind11/numpy.h>
-#include <pybind11/stl.h>
-#include <pybind11/functional.h>
-
-namespace py = pybind11;
-
-template<typename dtype>
-using np_array = py::array_t<dtype, py::array::c_style>;
