+++ sktime/src/README.md
@@ -1,0 +1,1 @@
+This directory contains common utility functionality for the c++ extensions contained in scikit-time.
